moved session related code to the SessionManagement namespace
