+++ src/Cassandra.IntegrationTests/Core/ReconnectionTests.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading;
 using Cassandra.IntegrationTests.TestBase;
 using Cassandra.IntegrationTests.TestClusterManagement;
+using Cassandra.SessionManagement;
 using Cassandra.Tests;
 using NUnit.Framework;
