+++ src/Cassandra.IntegrationTests/Core/SessionTests.cs
@@ -24,6 +24,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Cassandra.SessionManagement;
 using Cassandra.Tasks;
 
 namespace Cassandra.IntegrationTests.Core
