+++ src/Cassandra/Session.cs
@@ -24,6 +24,7 @@ using System.Threading.Tasks;
 
 using Cassandra.Requests;
 using Cassandra.Serialization;
+using Cassandra.SessionManagement;
 using Cassandra.Tasks;
 
 namespace Cassandra
