+++ src/Cassandra/SessionManagement/IInternalSession.cs
@@ -14,12 +14,12 @@
 //   limitations under the License.
 //
 
-namespace Cassandra
-{
-    using System.Collections.Generic;
-    using System.Net;
-    using System.Threading.Tasks;
+using System.Collections.Generic;
+using System.Net;
+using System.Threading.Tasks;
 
+namespace Cassandra.SessionManagement
+{
     /// <inheritdoc />
     /// <remarks>This is an internal interface designed to declare the internal methods that are called
     /// across multiple locations of the driver's source code.</remarks>
