+++ src/Cassandra/SessionState.cs
@@ -17,6 +17,7 @@
 using System.Collections.Generic;
 using System.Text;
 using Cassandra.Collections;
+using Cassandra.SessionManagement;
 
 namespace Cassandra
 {
